requirements.txt
torch>=2.0.0
transformers>=4.30.0
bitsandbytes>=0.40.0
peft>=0.4.0
datasets>=2.12.0
accelerate>=0.20.0
scikit-learn>=1.0.0
pyyaml>=6.0.0